+++ src/Business/Entities/Team.java
@@ -6,7 +6,7 @@ public class Team {
     private String name;
     private Member[] members;
 
-    public Team(String name, int id1, int id2, int id3, int id4) {
+    public Team(String name, long id1, long id2, long id3, long id4) {
         this.name = name;
         this.members = new Member[4];
         this.members[0] = new Member(id1, "Balanced");
